Extract Format into image, simplify
